Initial entry of the xmlSchemaParser and the xmlSchemaOwlPrinter.
The xmlSchemaParser reads an XML schema file, checks its syntax, and reprints
it with the suffix "echo" added. There are options for (1) keeping or
removing comments, and (2) reformatting or deleting documentation nodes.
The parser echos the file to the monitor while it parses. If there is a
parse error, echoing stops at that point and an error message is printed.

The xmlSchemaOwlPrinter writes an owl class file equivalent to an XML schema.
It can handle only a restricted set of constructs that may appear in an
XML schema file. It handles everything in kitting.xsd.


Former-commit-id: 9ac9ffe8bce450d7710fd8df5b317029018d8441