Massive changes to file structure and controller flow.
Made controller work with cmake to get rid of bizarre build requirements.
Moved many files to different areas to make directory structure more consistant.
Changed flow of controller to have queues of commands and status returns.
Added mutex to allow process to block when waiting for command or status return.
Seems to work for reading a file, not so sure about command line input.
Not sure why at this point, but wanted to get preliminary code checked in.

Signed-off-by: Stephen Balakirsky <stepen.balakirsky@gtri.gatech.edu>

Former-commit-id: 33e383374346bc287fa1dc36c5b2d5c801f0138c