Added ability to read files and have different joint and cart values.
Can now read CRCL command files where each command in the
file is sent after the previous commands completes with a
done status.

Separated the Cartesian and Joint max velocities and
accelerations so that they can be different for the
CRCL joint and cartesian moves.

Still waiting on NIST to provide feedback on the validity of
having joint moves in the CRCL.

Signed-off-by: Stephen Balakirsky <stepen.balakirsky@gtri.gatech.edu>